Move open day period to test source CIRC-280
