+++ src/test/java/api/support/OpeningDayPeriod.java
@@ -1,15 +1,13 @@
-package org.folio.circulation.domain;
+package api.support;
+
+import org.folio.circulation.domain.OpeningDay;
 
 import io.vertx.core.json.JsonObject;
 
 public class OpeningDayPeriod {
   private static final String OPENING_DAY_KEY = "openingDay";
 
-  private OpeningDay openingDay;
-
-  OpeningDayPeriod(JsonObject jsonObject) {
-    this.openingDay = new OpeningDay(jsonObject, OPENING_DAY_KEY);
-  }
+  private final OpeningDay openingDay;
 
   private OpeningDayPeriod(OpeningDay openingDay) {
     this.openingDay = openingDay;
@@ -18,7 +16,6 @@ public class OpeningDayPeriod {
   public static OpeningDayPeriod createDayPeriod(OpeningDay openingDay) {
     return new OpeningDayPeriod(openingDay);
   }
-
 
   public OpeningDay getOpeningDay() {
     return openingDay;
